Palette file rename
