+++ src/theme/index.ts
@@ -1,1 +1,1 @@
-export { default as Theme } from './theme';
+export { default as Theme } from './palette'
