+++ src/theme/palette.tsx
@@ -1,8 +1,8 @@
-import { createTheme } from '@mui/material/styles';
+import { createTheme } from '@mui/material/styles'
 
 const palette = {
   primary: {
-    main: '#1976d2', 
+    main: '#1976d2',
     contrastText: '#ffffff',
   },
   secondary: {
@@ -10,29 +10,28 @@ const palette = {
     contrastText: '#ffffff',
   },
   background: {
-    default: '#f4f6f8', 
-    paper: '#ffffff', 
+    default: '#f4f6f8',
+    paper: '#ffffff',
   },
   error: {
-    main: '#f44336', 
+    main: '#f44336',
   },
   warning: {
-    main: '#ff9800', 
+    main: '#ff9800',
   },
   info: {
-    main: '#2196f3', 
+    main: '#2196f3',
   },
   success: {
-    main: '#4caf50', 
+    main: '#4caf50',
   },
   text: {
-    primary: '#333333', 
-    secondary: '#757575', 
-    disabled: '#bdbdbd', 
+    primary: '#333333',
+    secondary: '#757575',
+    disabled: '#bdbdbd',
   },
-  divider: '#e0e0e0', 
-};
-
+  divider: '#e0e0e0',
+}
 
 const theme = createTheme({
   palette,
@@ -49,9 +48,9 @@ const theme = createTheme({
     // You can define other typography styles as well
   },
   shape: {
-    borderRadius: 8, 
+    borderRadius: 8,
   },
   // You can add more customizations here (e.g., spacing, breakpoints, etc.)
-});
+})
 
-export default theme;
+export default theme
